+++ src/streamlink/logger.py
@@ -5,7 +5,27 @@ from logging import CRITICAL, DEBUG, ERROR, INFO, WARNING
 from pathlib import Path
 from sys import version_info
 from threading import Lock
-from typing import IO, List, Optional, TYPE_CHECKING, Union
+from typing import IO, Iterator, List, Optional, TYPE_CHECKING, Union
+
+
+if TYPE_CHECKING:  # pragma: no cover
+    _BaseLoggerClass = logging.Logger
+else:
+    _BaseLoggerClass = logging.getLoggerClass()
+
+
+class StreamlinkLogger(_BaseLoggerClass):
+    def iter(self, level: int, messages: Iterator[str], *args, **kwargs) -> Iterator[str]:
+        """
+        Iterator wrapper for logging multiple items in a single call and checking log level only once
+        """
+
+        if not self.isEnabledFor(level):
+            yield from messages
+
+        for message in messages:
+            self._log(level, message, args, **kwargs)
+            yield message
 
 
 FORMAT_STYLE = "{"
@@ -19,8 +39,9 @@ REMOVE_BASE = ["streamlink", "streamlink_cli"]
 #  then it continues with the parent logger, which means that if the level of `streamlink.logger.root` was set to "none" and
 #  its value NOTSET, then it would continue with `logging.root` whose default level is `logging.WARNING` (equal to 30).
 NONE = sys.maxsize
-# Add "trace" to Streamlink's log levels
+# Add "trace" and "all" to Streamlink's log levels
 TRACE = 5
+ALL = 2
 
 # Define Streamlink's log levels (and register both lowercase and uppercase names)
 _levelToNames = {
@@ -31,34 +52,39 @@ _levelToNames = {
     INFO: "info",
     DEBUG: "debug",
     TRACE: "trace",
+    ALL: "all",
 }
+
+_custom_levels = TRACE, ALL
+
+
+def _logmethodfactory(level: int, name: str):
+    # fix module name that gets read from the call stack in the logging module
+    # https://github.com/python/cpython/commit/5ca6d7469be53960843df39bb900e9c3359f127f
+    if version_info >= (3, 11):
+        def method(self, message, *args, **kws):
+            if self.isEnabledFor(level):
+                # increase the stacklevel by one and skip the `trace()` call here
+                kws["stacklevel"] = 2
+                self._log(level, message, args, **kws)
+    else:
+        def method(self, message, *args, **kws):
+            if self.isEnabledFor(level):
+                self._log(level, message, args, **kws)
+
+    method.__name__ = name
+    return method
+
 
 for _level, _name in _levelToNames.items():
     logging.addLevelName(_level, _name.upper())
     logging.addLevelName(_level, _name)
 
+    if _level in _custom_levels:
+        setattr(StreamlinkLogger, _name, _logmethodfactory(_level, _name))
+
+
 _config_lock = Lock()
-
-
-if TYPE_CHECKING:  # pragma: no cover
-    _BaseLoggerClass = logging.Logger
-else:
-    _BaseLoggerClass = logging.getLoggerClass()
-
-
-class StreamlinkLogger(_BaseLoggerClass):
-    # fix module name that gets read from the call stack in the logging module
-    # https://github.com/python/cpython/commit/5ca6d7469be53960843df39bb900e9c3359f127f
-    if version_info >= (3, 11):
-        def trace(self, message, *args, **kws):
-            if self.isEnabledFor(TRACE):
-                # increase the stacklevel by one and skip the `trace()` call here
-                kws["stacklevel"] = 2
-                self._log(TRACE, message, args, **kws)
-    else:
-        def trace(self, message, *args, **kws):
-            if self.isEnabledFor(TRACE):
-                self._log(TRACE, message, args, **kws)
 
 
 class StringFormatter(logging.Formatter):
@@ -137,6 +163,7 @@ levels = list(_levelToNames.values())
 __all__ = [
     "NONE",
     "TRACE",
+    "ALL",
     "StreamlinkLogger",
     "basicConfig",
     "root",