+++ src/streamlink/plugins/twitch.py
@@ -2,7 +2,9 @@
 $description Global live-streaming and video hosting social platform owned by Amazon.
 $url twitch.tv
 $type live, vod
-$notes :ref:`Low latency streaming <cli/plugins/twitch:Low latency streaming>` is supported
+$notes See the :ref:`Authentication <cli/plugins/twitch:Authentication>` docs on how to prevent ads.
+$notes Read more about :ref:`embedded ads <cli/plugins/twitch:Embedded ads>` here.
+$notes :ref:`Low latency streaming <cli/plugins/twitch:Low latency streaming>` is supported.
 """
 
 import argparse
@@ -10,7 +12,7 @@ import json
 import logging
 import re
 import sys
-from datetime import datetime
+from datetime import datetime, timedelta
 from random import random
 from typing import List, NamedTuple, Optional
 from urllib.parse import urlparse
@@ -19,7 +21,7 @@ from streamlink.exceptions import NoStreamsError, PluginError
 from streamlink.plugin import Plugin, pluginargument, pluginmatcher
 from streamlink.plugin.api import validate
 from streamlink.stream.hls import HLSStream, HLSStreamReader, HLSStreamWorker, HLSStreamWriter
-from streamlink.stream.hls_playlist import ByteRange, ExtInf, Key, M3U8, M3U8Parser, Map, load as load_hls_playlist
+from streamlink.stream.hls_playlist import ByteRange, DateRange, ExtInf, Key, M3U8, M3U8Parser, Map, load as load_hls_playlist
 from streamlink.stream.http import HTTPStream
 from streamlink.utils.args import keyvalue
 from streamlink.utils.parse import parse_json, parse_qsd
@@ -70,27 +72,31 @@ class TwitchM3U8Parser(M3U8Parser):
         # This is better than using the duration of the last segment when regular segment durations vary a lot.
         # In low latency mode, the playlist reload time is the duration of the last segment.
         duration = last.duration if last.prefetch else sum(segment.duration for segment in segments) / float(len(segments))
-        segments.append(last._replace(
+        # Use the last duration for extrapolating the start time of the prefetch segment, which is needed for checking
+        # whether it is an ad segment and matches the parsed date ranges or not
+        date = last.date + timedelta(seconds=last.duration)
+        ad = self._is_segment_ad(date)
+        segment = last._replace(
             uri=self.uri(value),
             duration=duration,
-            prefetch=True
-        ))
+            title=None,
+            discontinuity=self.state.pop("discontinuity", False),
+            date=date,
+            ad=ad,
+            prefetch=True,
+        )
+        segments.append(segment)
 
     def parse_tag_ext_x_daterange(self, value):
         super().parse_tag_ext_x_daterange(value)
         daterange = self.m3u8.dateranges[-1]
-        is_ad = (
-            daterange.classname == "twitch-stitched-ad"
-            or str(daterange.id or "").startswith("stitched-ad-")
-            or any(attr_key.startswith("X-TV-TWITCH-AD-") for attr_key in daterange.x.keys())
-        )
-        if is_ad:
+        if self._is_daterange_ad(daterange):
             self.m3u8.dateranges_ads.append(daterange)
 
     def get_segment(self, uri: str) -> TwitchSegment:  # type: ignore[override]
         extinf: ExtInf = self.state.pop("extinf", None) or ExtInf(0, None)
         date = self.state.pop("date", None)
-        ad = any(self.m3u8.is_date_in_daterange(date, daterange) for daterange in self.m3u8.dateranges_ads)
+        ad = self._is_segment_ad(date, extinf.title)
 
         return TwitchSegment(
             uri=uri,
@@ -102,7 +108,21 @@ class TwitchM3U8Parser(M3U8Parser):
             date=date,
             map=self.state.get("map"),
             ad=ad,
-            prefetch=False
+            prefetch=False,
+        )
+
+    def _is_segment_ad(self, date: datetime, title: Optional[str] = None) -> bool:
+        return (
+            title is not None and "Amazon" in title
+            or any(self.m3u8.is_date_in_daterange(date, daterange) for daterange in self.m3u8.dateranges_ads)
+        )
+
+    @staticmethod
+    def _is_daterange_ad(daterange: DateRange) -> bool:
+        return (
+            daterange.classname == "twitch-stitched-ad"
+            or str(daterange.id or "").startswith("stitched-ad-")
+            or any(attr_key.startswith("X-TV-TWITCH-AD-") for attr_key in daterange.x.keys())
         )
 
 
@@ -234,6 +254,8 @@ class TwitchAPI:
             "Client-ID": "kimne78kx3ncx6brgo4mv6wki5h1ko",
         }
         self.headers.update(**dict(session.get_plugin_option("twitch", "api-header") or []))
+        self.access_token_params = dict(session.get_plugin_option("twitch", "access-token-param") or [])
+        self.access_token_params.setdefault("playerType", "embed")
 
     def call(self, data, schema=None):
         res = self.session.http.post(
@@ -386,7 +408,7 @@ class TwitchAPI:
             login=channel_or_vod if is_live else "",
             isVod=not is_live,
             vodID=channel_or_vod if not is_live else "",
-            playerType="embed"
+            **self.access_token_params,
         )
         subschema = validate.none_or_all(
             {
@@ -513,6 +535,19 @@ class TwitchAPI:
         A header to add to each Twitch API HTTP request.
 
         Can be repeated to add multiple headers.
+
+        Useful for adding authentication data that can prevent ads. See the plugin-specific documentation for more information.
+    """,
+)
+@pluginargument(
+    "access-token-param",
+    metavar="KEY=VALUE",
+    type=keyvalue,
+    action="append",
+    help="""
+        A parameter to add to the API request for acquiring the streaming access token.
+
+        Can be repeated to add multiple parameters.
     """,
 )
 class Twitch(Plugin):